% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_supergraph.R
\name{cliqueSimilarity}
\alias{cliqueSimilarity}
\title{Compute percentages of all possible connections between cliques}
\usage{
cliqueSimilarity(cl_mem, WorA, do_signif = TRUE, lvl = 0.05)
}
\arguments{
\item{cl_mem}{an integer vector of nonnegative clique labels where 0 encodes sssumed group of nodes outside of all cliques. Length of \code{cl_mem} must be equal to number of rows/columns of \code{worA}.}

\item{WorA}{An adjacency or weight matrix of a graph in which the cliques encoded by \code{cl_mem} were found.}

\item{do_signif, lvl}{If \code{do_signif} is \code{TRUE}, zero out connections between cliques below statistical significance threshold of \code{lvl} after multiple test correction, see details}
}
\value{
A symmetric matrix giving clique to clique similarity, optionally with an additional attribute "labelMap", see details.
}
\description{
Given a weight/adjacency matrix \code{A} and a membership vector the function counts the present edges between
each pair of cliques and divides it by maximum possible number of such edges. Result of the calculation
is stored in a symmetric matrix: \code{cS[i,j]} is the percentage of possible connections between clique number \code{i}
and \code{j}.
}
\details{
Function assumes groups are cliques, so diagonal of the resulting matrix is set to 1 (all possible edges inside
each group are present).

\code{WorA} is always converted into binary adjacency (1 if entry \verb{>0}).

If there are \code{N} unique non zero labels in \code{cl_mem}, the resulting matrix is of size \verb{N x N}.
Labels in \code{cl_mem} might get converted to consecutive integers by \code{\link[=tidyUpLabels]{tidyUpLabels()}} and based on the result of
such conversion the mapping from clique label to column/row number is defined.
If sorted unique non zero labels in \code{cl_mem} were not exactly equal to \code{1:max(cl_mem)}, the
mapping from input to output labels is saved in a form of a 2 data.frame in an attribute
"labelMap". 1st column contains new consecutive labels and second column the previous, original labels.

Note that if cliques of \code{cl_mem} were found in a thresholded version of the similarity matrix, input \code{WorA} should be the thresholded version.
}
\examples{
data(leukemia)
opt_<- thr_optimizer(leukemia)
cliqueSimilarity(opt_$maximizer_partition, leukemia \%thr\% opt_$thr)-> cS
dim(cS)
table(opt_$maximizer_partition) 
}
\seealso{
\code{\link[cliquePartitioneR:tidyUpLabels]{cliquePartitioneR::tidyUpLabels()}}, \code{\link[cliquePartitioneR:greedyCliquePartitioner]{cliquePartitioneR::greedyCliquePartitioner()}}, \code{\link[=\%thr\%]{\%thr\%()}}
}
