% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operators.R
\name{flip_}
\alias{flip_}
\title{Convert simmilarities to dissimilarities and vice versa}
\usage{
flip_(W)
}
\arguments{
\item{W}{Numeric matrix}
}
\value{
A numeric matrix "flipped" over its maximum
}
\description{
Each entry of matrix \code{M[i,j]} in the output gets replaced by \code{max(M) - M[i,j]}
}
\examples{
library(magrittr)
data(leukemia)
leukemia[1:4,1:4]
leukemia[1:4,1:4] \%>\% flip_()
}
