% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operators.R
\name{\%ideg\%}
\alias{\%ideg\%}
\title{Calculate inner degree of the cluster in network based ow weight matrix}
\usage{
W \%ideg\% mask
}
\arguments{
\item{W}{Numeric matrix}

\item{mask}{logical mask}
}
\value{
\code{sum(W[mask, mask])}
}
\description{
Calculate inner degree of the cluster in network based ow weight matrix
}
