% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operators.R
\name{\%odeg\%}
\alias{\%odeg\%}
\title{Calculate outer degree of the cluster in network based ow weight matrix}
\usage{
W \%odeg\% mask
}
\arguments{
\item{W}{Numeric matrix}

\item{mask}{logical mask}
}
\value{
\code{sum(W[mask, !mask])}
}
\description{
Calculate outer degree of the cluster in network based ow weight matrix
}
