% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operators.R
\name{\%[\%}
\alias{\%[\%}
\title{Index both rows and columns of a matrix}
\usage{
W \%[\% idx
}
\arguments{
\item{W}{Numeric matrix}

\item{idx}{vector of indexes or logical mask}
}
\value{
\code{W[idx,idx]}
}
\description{
Index both rows and columns of a matrix
}
