% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_utils.R
\name{\%subdivides\%}
\alias{\%subdivides\%}
\title{Check if partition Q subdivides partition P}
\usage{
Q \%subdivides\% P
}
\arguments{
\item{Q}{membership vector}

\item{P}{membership vector}
}
\value{
\code{TRUE} if \code{Q} subdivides \code{P}, \code{FALSE} if there are overlaps
}
\description{
Check if partition Q subdivides partition P
}
