% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operators.R
\name{\%thr\%}
\alias{\%thr\%}
\title{Zero out weights in similarity matrix less than specified threshold and return thresholded matrix}
\usage{
W \%thr\% t
}
\arguments{
\item{W}{Weight matrix to be thresholded}

\item{t}{A value below which weights will be zeroed out}
}
\value{
A weight matrix \code{W} but with weights below \code{t} set to zero
}
\description{
Convenience function for thresholding of similarity matrix in simplified syntax
}
\examples{
M<-matrix(runif(100),nrow=10)
print(min(M))
M \%thr\% 0.5 -> M
print(min(M))
}
