% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operators.R
\name{igr_}
\alias{igr_}
\title{Convert symmetric matrix to undirected weighted graph from igraph}
\usage{
igr_(W)
}
\arguments{
\item{W}{Weight matrix of the graph to construct}
}
\value{
Undirected weighted graph object from igraph with weights from \code{W}
}
\description{
Note: this discards loops (diagonal of the matrix)
}
\examples{
data(leukemia)
hist(igraph::transitivity(igr_(W), type="weighted"), main="weighted clustering coefficient")
}
