% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clique_cluste.R
\name{MinST_DunnIndex}
\alias{MinST_DunnIndex}
\title{Calculate Dunn Index goodness of clustering variant based on Minimum Spanning Tree}
\usage{
MinST_DunnIndex(partition, dS, outlier_mask = NULL)
}
\arguments{
\item{partition}{A vector of membership labels of clusters of each node, 0 labels get converted to singleton labels by \code{\link[cliquePartitioneR:uniqueSingletonLabels]{cliquePartitioneR::uniqueSingletonLabels()}}}

\item{dS}{A dissimilarity matrix of nodes in the graph. The higher the \code{dS[i,j]} the more dissimilar nodes \code{i} and \code{j} are.}

\item{outlier_mask}{(experimental) if not null, if \code{outlier_mask[i]} is \code{TRUE} then node \code{i} is not included in the calculation, see details.}
}
\value{
A scalar value indicating the goodness of clustering, the bigger the better. Can be \code{Inf} if max cluster diameter is 0. If both \code{min_between_cluster_distance} and \code{max_cluster_diameter} are 0, function returns 0 by convention.
}
\description{
For a given dissimilarity graph \code{dS} and partition of \code{dS} into
clusters, function will compute Dunn Index considering
dissimilarities of Minimum Spanning Tree (MST), see details.
}
\details{
In general, Dunn Index is defined as \code{min_between_cluster_distance/max_cluster_diameter} - the bigger
the value, the better the clustering.

Note that it is formulated in terms of distances, not in similarities, so to keep interpretation consistent
with literature this function works with dissimilarities. Below by distance we mean dissimilarity.
Dissimilarity can be derived from similarity matrix for example by \code{dS= max(S) - S}.

This function computes \code{between_cluster_distance} between clusters \verb{i,j} as a minimum distance between
any nodes inside clusters \verb{i,j}.
\code{cluster_diameter} is defined as maximal nearest-neighbor distance in the cluster.
Both of those quantities are derived from minimum spanning tree of the graph described by \code{dS}.

Value of the index is undefined if there is only one cluster or all clusters are singletons, in that case
\code{NA} is returned.
For the case when maximum cluster diameter is 0 value of the index is infinity
and if also minimum between cluster distance is 0 then the function returns 0 (there are no differences
between inter and intra cluster distances so a given clustering is not correct).

If \code{outlier_mask} is not \code{NULL}, then it is assumed that it is a logcial vector of length equal to number of nodes in the network.
If \code{outlier_mask[[i]]} is \code{TRUE} then node \code{i} gets removed from the network before calculating Dunn Index.
This is an ad hoc method of masking some of the particularly problematic nodes that connect clusters
by very few but strong connections (lowering the \code{between_cluster_distance}).
Criteria for being an outlier here can be based on low clustering coefficient (transitivity), see \code{\link[igraph:transitivity]{igraph::transitivity()}}.

Note that even though package offers various thresholding schemes for similarity matrices, for assesing
goodness of the clustering one should calculate this based on dissimlarity derived from the unthresholded version
of the graph.
}
