% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_scpcss}
\alias{run_scpcss}
\title{Run Simple Clique Paritioning based Clustering of Similiarty Spaces}
\usage{
run_scpcss(
  XorS,
  t_S = "complexity",
  t_CS = "modularity",
  X_simil_fun = NULL,
  return_S = FALSE,
  return_CS = FALSE,
  expansion_mode = "basic",
  t_S_init_steps = 10,
  t_S_keep_init_eval_history = TRUE,
  t_S_keep_init_partitions = FALSE,
  t_S_precis = TRUE,
  t_S_tol = 0.1,
  t_CS_relax_method = "greedyCliqueJoin",
  t_CS_cs_thr_objective = "modularity",
  t_CS_n_init_steps = 10,
  t_CS_keep_init_eval_history = TRUE,
  t_CS_keep_init_partitions = FALSE,
  t_CS_precis = TRUE,
  t_CS_tol = 0.1
)
}
\arguments{
\item{XorS}{Either a matrix where each row gives coordinates of points between which to compute similarities or node similarity matrix, see details.}

\item{t_S}{A threshold choice for similarity matrix. Either character string determining strategy to choose this value optimally, a numeric value giving exact thr. to use or a function computing this thr. based on similarity matrix \code{S}, see details.}

\item{t_CS}{A threshold choice of clique similarity matrix used in clique relaxation step. Either a character string specifying strategy to use for finding this value or an exact numeric value to use, see details.}

\item{X_simil_fun}{If not \code{NULL}, then it is assumed \code{XorS} is a matrix of coordinates of points (one row per one point). If a character string, then it specifies distance metric to use for computing similarities. If a function, then it will be used  on \code{XorS} to produce similarity matrix for clique partitioning.}

\item{return_S, return_CS}{Logical values, if \code{TRUE}, then return similarity matrix \code{S} and clique similarity matrix \code{CS} utilized, respectively.}

\item{expansion_mode}{A clique partitioning strategy to use, see \code{\link[cliquePartitioneR:greedyCliquePartitioner]{cliquePartitioneR::greedyCliquePartitioner()}}.}

\item{t_S_init_steps, t_S_keep_init_eval_history, t_S_keep_init_partitions, t_S_precis, t_S_tol}{Passed to \code{\link[=thr_optimizer]{thr_optimizer()}} if t_S is set to one of \verb{"complexity","n_cliques"}. These parameters control threshold optimization process for similarity matrix \code{S}, see \code{\link[=thr_optimizer]{thr_optimizer()}} for in depth description.}

\item{t_CS_relax_method}{Character string specifying clique relaxation method to use, see \code{\link[=relax_cliques]{relax_cliques()}}.}

\item{t_CS_cs_thr_objective, t_CS_n_init_steps, t_CS_keep_init_eval_history, t_CS_keep_init_partitions, t_CS_precis, t_CS_tol}{Passed to \code{\link[=relax_cliques_optimally]{relax_cliques_optimally()}} if \code{t_CS} is set to one of \verb{"mst","modularity"}. These parameters control threshold optimization process for clique similarity matrix \code{CS}, see \code{\link[=relax_cliques_optimally]{relax_cliques_optimally()}} for in depth description.}
}
\value{
A named list with the following components:
\itemize{
\item \code{clique_membership} A clique membership vector of nodes. Singletons are given unique labels.
\item \code{t_S} A numerical value of threshold used on similarity matrix \code{S} before clique partitioning.
\item \code{t_CS} A numerical value of threshold used on clique similarity matrix in clique relaxation process.
\item \code{cluster_membership} A sublist of 2 elements, each giving a kind of membership vector:
\itemize{
\code{cliq} gives cluster labes to which cliques belong
\code{node} gives cluster labels to which nodes belong (cluster of node \code{i} is the cluster of clique in which \code{i} is situated)
}
\item \code{S} A symmetic matrix of node similarities used (non thresholded, set if \code{return_S} was \code{TRUE})
\item \code{t_S_objective} A value of objective function obtained for final used \code{t_S} (set if threshold optimization of matrix \code{S} was done)
\item \code{t_S_init_partitions} Partitions corresponing to inital search for \code{t_S} (set if threshold optimization of matrix \code{S} was done and \code{t_S_keep_init_partitions} was \code{TRUE})
\item \code{t_S_init_search_points} Initial search points for \code{t_S}  (set if inital threshold optimization of matrix \code{S} was done and \code{t_S_keep_init_score} was \code{TRUE})
\item \code{t_S_init_scores} Initial scores of each initial threshold in \code{t_S_init_search_points}  (set if inital threshold optimization of matrix \code{S} was done and \code{t_S_keep_init_score} was \code{TRUE})
\item \code{CS} A symmetic matrix of clique similarities used (non thresholded, set if \code{return_CS} was \code{TRUE})
\item \code{t_CS_objective} A value of objective function obtained for final used \code{t_CS} (set if threshold optimization of matrix \code{CS} was done)
\item \code{t_CS_init_partitions} Partitions corresponing to inital search for \code{t_CS} (set if initial threshold optimization of matrix \code{CS} was done and \code{t_CS_keep_init_partitions} was \code{TRUE})
\item \code{t_CS_init_scores} Initial scores of each initial threshold in \code{t_CS_init_search_points}  (set if inital threshold optimization of matrix \code{CS} was done and \code{t_CS_keep_init_score} was \code{TRUE})
}
}
\description{
Given an input matrix of pairwise node similarities or coordinates of objects in some metric space \code{XorS},
function will cluster the nodes/points based on their similarity matrix by
partitioning similarity graph into cliques  and relaxing the cliques into clusters, both (optionally) in an automatic,
optmial way.
}
